sample_processes/pyproject.toml
[tool.poetry]
name = "sample-processes"
version = "0.1.0"
description = ""
authors = ["a.petrov"]
readme = "README.md"
packages = [{include = "sample_processes"}]

[tool.poetry.dependencies]
python = "^3.10"
pykafka = "^2.8.0"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
pyopenssl = "^23.1.1"
loguru = "^0.7.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.262"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
